Segregate based on usage
Rename `MagicSystem` to `MagicUI`
Move cache stuff into `MagicCache`
Move `_productToRecipes` to `MagicCache` as `ResultToRecipe`
